refactor(react-components): ディレクトリ名を複数形に統一
src/primitive/ → src/primitives/、src/view/ → src/views/、
src/lib/ → src/libs/ に揃え、他パッケージ（features / helpers /
schemas / testing 等の複数形）と命名規則を統一する。

ADR-024 が決定した「層」の概念（プリミティブ層 / ビュー層）は
維持したまま、ディレクトリ名だけを複数形にする扱い。

破壊的変更:
- @next-lift/react-components の subpath export `/lib` を `/libs`
  に変更（apps/web 側の利用箇所も合わせて更新）

src/demo/ は据え置き（playground 残置物として将来削除予定のため
リネームコストをかけない）。
