+++ packages/react-components/src/primitives/button/button.tsx
@@ -6,7 +6,7 @@ import {
 	type ButtonProps as ButtonPrimitiveProps,
 } from "react-aria-components";
 import { tv, type VariantProps } from "tailwind-variants";
-import { cx } from "../../lib/primitive";
+import { cx } from "../../libs/primitive";
 
 export const buttonStyles = tv({
 	base: [
