+++ packages/react-components/src/primitives/modal/dialog.tsx
@@ -8,7 +8,7 @@ import {
 	Button as PrimitiveButton,
 	Dialog as PrimitiveDialog,
 } from "react-aria-components";
-import { cn } from "../../lib/utils";
+import { cn } from "../../libs/utils";
 import { Button } from "../button";
 
 type DialogProps = {
